Create hw3-The Honest Student
Instructions
The god of NCCU will ask you if you dropped the golden card, silver card, or just an ordinary student card.

Input your answer by character in the program. 'G' means golden card, 'S' means silver card, and 'O' means ordinary student card.

And also input an integer N which is the money you have in your card.

There are three conditions will happen then:

1. If you are honest, he will say that you are a "honest student" and deposit some money in it. The amount will be the sum of all the factors of N.

For example, you have $1000 in your card before you dropped it.
The factors of 1000 are {1, 2, 4, 5, 8, 10, 20, 25, 40, 50, 100, 125, 200, 250, 500, 1000}.
And the sum of all the factors is 2340, which means the god deposit $2340 in it.
Then you will get $1000 + $2340 = $3340 in your card now.
2. If you tell a lie, he will say that you are a "liar" and will not give you any money or even your card back.

3. If you say any answer which he is unexpecting to hear (which means the input without 'G', 'S', or 'O'), he will say "???" and just return you the card without any reward.

Input Formats
The first line of the input consists of an character indicating your answer to the god.

The input character can be any English character from 'A' to 'Z'. And noted that the character is case-insensitive.

The second line of the input consists of an integer N (0 <= N <= 30000) indicating the money you have in your student card before you dropped it into the river.

Output Formats
The first line is the answer of the God.

The second line is the money you have in your student card after the god returned it to you.

Noted that there will not have second line if you lie to the god.

Please see the sample below for the right format.

Input Sample 1
O

1000

Output Sample 1
Honest student.

You have $3340 in your card now.

Input Sample 2
G

1000

Output Sample 2
Liar. This golden card is not yours.

Input Sample 3
A

1000

Output Sample 3
???

You have $1000 in your card now.